renamed helper files for more consistency btwn python and javascript helpers
