+++ translate/assign/js_assign.py
@@ -4,7 +4,7 @@ import os
 # get path automatically and handle imports
 path = os.path.dirname(os.getcwd()) + '/helpers'
 sys.path.insert(1, path)
-import steph_helpers
+import js_helpers
 
 """
 Handles js var declerations to generic AST node
@@ -15,5 +15,5 @@ def jsassign_to_gast(node):
 
   # only works with single assignment
   gast["varId"] = node[0].id.name
-  gast["varValue"] = steph_helpers.jsarg_to_str(node[0].init)
+  gast["varValue"] = js_helpers.jsarg_to_str(node[0].init)
   return gast
