+++ translate/helpers/js_helpers.py
