+++ translate/py_main.py
@@ -1,7 +1,7 @@
 from routers.py_router import *
 from expression.py_expression import *
 from assign.py_assign import *
-from helpers.jack_py_helpers import *
+from helpers.py_helpers import *
 import ast
 import astor
